+++ pymutspec/__init__.py
@@ -1,4 +1,4 @@
 from pymutspec import annotation, draw, io, constants, utils
 
 # Define the pymutspec version
-__version__ = "0.0.11"
+__version__ = "0.0.12"
