+++ pyproject.toml
@@ -1,17 +1,39 @@
+[build-system]
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
+
 [project]
 name = "PyMutSpec"
-version = "0.0.11"
+version = "0.0.12"
+description = "Mutational spectra analysis package"
 authors = [
   { name="kpotoh", email="axepon@mail.ru" },
 ]
-description = "Mutational spectra analysis package"
 license = { file = "LICENSE" }
-readme = "README.md"
+readme = {file = "README.md", content-type = "text/markdown"}
 requires-python = ">=3.8"
 classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
+    "Development Status :: 3 - Alpha",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Topic :: Scientific/Engineering",
+    "Topic :: Scientific/Engineering :: Bio-Informatics",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "License :: OSI Approved :: MIT License",
+]
+dependencies = [
+    "numpy~=1.22.2",
+    "pandas~=1.4.0",
+    "seaborn==0.12.2",
+    "biopython==1.79",
+    "ete3==3.1.2",
+    "tqdm~=4.62.3",
+    "PyYAML~=6.0",
+    "pytest~=7.1.2",
+    "click~=8.1.3",
 ]
 
 [project.urls]
